Moved AtmosphereProcessType declaration and utils in a separate header.
