+++ components/scream/src/share/atmosphere_process.hpp
@@ -4,6 +4,7 @@
 #include <string>
 #include <set>
 
+#include "share/atmosphere_process_utils.hpp"
 #include "share/scream_assert.hpp"
 #include "share/mpi/scream_comm.hpp"
 #include "share/field/field_identifier.hpp"
@@ -15,25 +16,6 @@
 
 namespace scream
 {
-
-enum class AtmosphereProcessType {
-  Coupling,   // Process responsible of interfacing with the component coupler
-  Dynamics,   // Process responsible of handling the dynamics
-  Physics,    // Process handling a physics parametrization
-  Group       // Process that groups a bunch of processes (so they look as a single process)
-};
-
-inline std::string e2str (const AtmosphereProcessType ap_type) {
-  switch (ap_type) {
-    case AtmosphereProcessType::Coupling:  return "Surface Coupling";
-    case AtmosphereProcessType::Dynamics:  return "Atmosphere Dynamics";
-    case AtmosphereProcessType::Physics:   return "Atmosphere Physics Parametrization";
-    case AtmosphereProcessType::Group:     return "Atmosphere Process Group";
-    default:
-      error::runtime_abort("Error! Unrecognized atmosphere process type.\n");
-  }
-  return "INVALID";
-}
 
 /*
  *  The abstract interface of a process of the atmosphere 
@@ -114,6 +96,7 @@ protected:
   virtual void set_computed_field_impl (const Field<      Real, device_type>& f) = 0;
 };
 
+// A short name for the factory for atmosphere processes
 using AtmosphereProcessFactory = util::Factory<AtmosphereProcess,util::CaseInsensitiveString,const ParameterList&>;
 
 } // namespace scream
